estudo/NetTopologyScanner/NetTopologyScanner/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "mac-vendor-lookup>=0.1.12",
    "netaddr>=1.3.0",
    "netifaces>=0.11.0",
    "python-nmap>=0.7.1",
    "scapy>=2.6.1",
    "tabulate>=0.9.0",
]
